requirements.txt
flask==2.0.1
web3==5.24.0
python-dotenv==0.19.1
requests==2.26.0
eth-utils==1.10.0
eth-account==0.5.6
eth-typing==2.2.2
eth-hash==0.3.2
eth-abi==2.1.1
hexbytes==0.2.2
